requirements.txt
CouchDB==0.9
Flask==0.10.1
Flask-CouchDB==0.2.1
Flask-Mail==0.9.0
Flask-OAuthlib==0.4.3
Jinja2==2.7.2
MarkupSafe==0.19
PyYAML==3.11
Werkzeug==0.9.4
amqp==1.4.4
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.16
blinker==1.3
celery==3.1.10
flower==0.6.0
itsdangerous==0.24
kombu==3.0.14
mock==1.0.1
nose==1.3.0
oauthlib==0.6.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-termstyle==0.1.10
pytz==2014.2
rednose==0.4.1
sleekxmpp==1.2.4
sure==1.2.5
tornado==3.2
wsgiref==0.1.2
